commited to keep changes

+++ API Backend/src/main/java/org/example/futureengineers/Services/ServicesImp/ModelApiServiceImp.java
@@ -19,15 +19,15 @@ public class ModelApiServiceImp implements ModelApiService {
 
 
     @Override
-    public ModelApiRequestDto SendRequestAndFetchResult(Map<String, Float> requestDto) {
-        Map<String,Float> requestBody =requestDto;
+    public ModelApiRequestDto SendRequestAndFetchResult(Map<String, Integer> requestDto) {
+        Map<String,Integer> requestBody =requestDto;
 
         // Definir le request HTTP avec le type JSON
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.APPLICATION_JSON);
 
         // Construire la requête
-        HttpEntity<Map<String, Float>> requestEntity = new HttpEntity<>(requestBody, headers);
+        HttpEntity<Map<String, Integer>> requestEntity = new HttpEntity<>(requestBody, headers);
 
         // Envoyer la requête POST et récupérer la réponse {"best_major" : responce:string }
         ResponseEntity<ModelApiRequestDto> response = restTemplate.postForEntity(modelApiUrl, requestEntity, ModelApiRequestDto.class);
